Move AbstractStructuredGraphViewer to public API
